+++ src/NzbDrone.Core/MediaFiles/Commands/ConvertEbookCommand.cs
@@ -1,0 +1,13 @@
+using NzbDrone.Core.Messaging.Commands;
+
+namespace NzbDrone.Core.MediaFiles.Commands
+{
+    public class ConvertEbookCommand : Command
+    {
+        public int BookFileId { get; set; }
+
+        public override bool SendUpdatesToClient => true;
+        public override bool RequiresDiskAccess => true;
+        public override bool IsLongRunning => true;
+    }
+}